+++ crates/core/machine/src/control_flow/auipc.rs
@@ -130,7 +130,7 @@ impl<F: PrimeField32> MachineAir<F> for AuipcChip {
 
                         cols.state.populate(
                             &mut blu,
-                            input.public_values.execution_shard,
+                            input.public_values.execution_shard as u32,
                             event.0.clk,
                             event.0.pc,
                         );
