+++ crates/core/machine/src/memory/instructions/store/store_half.rs
@@ -104,7 +104,7 @@ impl<F: PrimeField32> MachineAir<F> for StoreHalfChip {
                         self.event_to_row(&event.0, cols, &mut blu);
                         cols.state.populate(
                             &mut blu,
-                            input.public_values.execution_shard,
+                            input.public_values.execution_shard as u32,
                             event.0.clk,
                             event.0.pc,
                         );
